refactor(nanorpc/server): move test mocks to centralized testutils_test.go
- Move mockSession implementations from individual test files to testutils_test.go
- Move newTestSession and newTestRequest helper functions to central location
- Move mockSessionWithError from request_context_test.go to testutils_test.go
- Fix EncodeResponse call signature in centralized implementation
- Establish single source of truth for server test utilities
