+++ pkg/nanorpc/server/handler_request_test.go
@@ -96,45 +96,3 @@ func TestDefaultMessageHandler_HandleRequest(t *testing.T) {
 		t.Run(tt.name, tt.test)
 	}
 }
-
-// mockSession implements Session interface for testing
-type mockSession struct {
-	id           string
-	remoteAddr   string
-	lastResponse *nanorpc.NanoRPCResponse
-	lastData     []byte
-}
-
-func (m *mockSession) ID() string {
-	return m.id
-}
-
-func (m *mockSession) RemoteAddr() string {
-	return m.remoteAddr
-}
-
-func (*mockSession) Handle(_ context.Context) error {
-	return nil
-}
-
-func (*mockSession) Close() error {
-	return nil
-}
-
-// SendResponse captures the response for testing
-func (m *mockSession) SendResponse(req *nanorpc.NanoRPCRequest, response *nanorpc.NanoRPCResponse) error {
-	// Fill envelope fields from request if provided
-	if req != nil && response.RequestId == 0 {
-		response.RequestId = req.RequestId
-	}
-
-	// Encode the response
-	data, err := nanorpc.EncodeResponse(response, nil)
-	if err != nil {
-		return err
-	}
-
-	m.lastData = data
-	m.lastResponse = response
-	return nil
-}
