+++ pkg/nanorpc/server/handler_test.go
@@ -10,34 +10,6 @@ import (
 
 // Test helpers and factories
 
-func newTestSession() *mockSession {
-	return &mockSession{
-		id:         "test-session",
-		remoteAddr: "127.0.0.1:12345",
-	}
-}
-
-func newTestRequest(id int32, pathOneOf any) *nanorpc.NanoRPCRequest {
-	req := &nanorpc.NanoRPCRequest{
-		RequestId:   id,
-		RequestType: nanorpc.NanoRPCRequest_TYPE_REQUEST,
-	}
-
-	switch p := pathOneOf.(type) {
-	case string:
-		req.PathOneof = nanorpc.GetPathOneOfString(p)
-	case uint32:
-		req.PathOneof = nanorpc.GetPathOneOfHash(p)
-	case nanorpc.PathOneOf:
-		req.PathOneof = p
-	default:
-		// This shouldn't happen in tests, but be defensive
-		req.PathOneof = nil
-	}
-
-	return req
-}
-
 func registerTestHandler(t testutils.T, handler *DefaultMessageHandler, path string, response []byte) {
 	t.Helper()
 	err := handler.RegisterHandlerFunc(path, func(_ context.Context, req *RequestContext) error {
